Feature/non linearity

+++ Connect-A-Pic-Core/ExternalPorts/LaserType.cs
@@ -29,7 +29,7 @@ namespace CAP_Core.ExternalPorts
             }
         }
         public LightColor Color { get; }
-        public double WaveLengthInNm { get; }
+        public int WaveLengthInNm { get; }
         public override string ToString()
         {
             return $"color: {Color.ToReadableString()}, WaveLength: {WaveLengthInNm}nm";
